requirements.txt
streamlit>=1.36
pandas>=2.2
google-cloud-bigquery>=3.18
nltk>=3.9
eng-to-ipa>=0.0.2
pronouncing>=0.2.0
setuptools>=70
